Move Module into mem_cache
