+++ .gitignore
@@ -5,7 +5,7 @@ __pycache__/
 
 # C extensions
 *.so
-.benchmark/
+
 # Distribution / packaging
 .Python
 build/
