+++ python/sglang/srt/layers/attention/nsa_backend.py
@@ -23,9 +23,7 @@ from sglang.srt.layers.attention.nsa.utils import (
     compute_nsa_seqlens,
 )
 from sglang.srt.layers.dp_attention import get_attention_tp_size
-from sglang.srt.mem_cache.allocator import (
-    is_enable_hierarchical_nsa,
-)
+from sglang.srt.mem_cache.allocator import is_enable_hierarchical_nsa
 from sglang.srt.mem_cache.memory_pool import NSAReqToTokenPool
 from sglang.srt.model_executor.forward_batch_info import ForwardBatch, ForwardMode
 from sglang.srt.utils import is_hip
