+++ python/sglang/srt/mem_cache/sparsity/algorithms/__init__.py
@@ -1,0 +1,15 @@
+from sglang.srt.mem_cache.sparsity.algorithms.base_algorithm import (
+    BaseSparseAlgorithm,
+    FakeRandomSparseAlgorithm,
+    SparseMode,
+)
+from sglang.srt.mem_cache.sparsity.algorithms.deepseek_nsa import DeepSeekNSAAlgorithm
+from sglang.srt.mem_cache.sparsity.algorithms.page_mean_pooling import PageMeanPoolingAlgorithm
+
+__all__ = [
+    "BaseSparseAlgorithm",
+    "SparseMode",
+    "FakeRandomSparseAlgorithm",
+    "PageMeanPoolingAlgorithm",
+    "DeepSeekNSAAlgorithm",
+]
