+++ python/sglang/srt/mem_cache/sparsity/algorithms/deepseek_nsa.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING, Any, Dict, Optional
 import nvtx
 import torch
 
-from sglang.srt.sparsity2.algorithms.base_algorithm import (
+from sglang.srt.mem_cache.sparsity.algorithms.base_algorithm import (
     BaseSparseAlgorithm,
     SparseMode,
 )
