--- python/sglang/srt/sparsity2/algorithms/__init__.py
@@ -1,15 +1,0 @@
-from sglang.srt.sparsity2.algorithms.base_algorithm import (
-    BaseSparseAlgorithm,
-    FakeRandomSparseAlgorithm,
-    SparseMode,
-)
-from sglang.srt.sparsity2.algorithms.deepseek_nsa import DeepSeekNSAAlgorithm
-from sglang.srt.sparsity2.algorithms.page_mean_pooling import PageMeanPoolingAlgorithm
-
-__all__ = [
-    "BaseSparseAlgorithm",
-    "SparseMode",
-    "FakeRandomSparseAlgorithm",
-    "PageMeanPoolingAlgorithm",
-    "DeepSeekNSAAlgorithm",
-]
